Fixed a bug where the counters in UserProfileResponse were int, but need to be long.

+++ SeawaveApp/Models/UserProfileResponse.cs
@@ -2,5 +2,10 @@ using System;
 
 namespace SeawaveApp.Models;
 
-public record UserProfileResponse(string Username, string Email, DateTimeOffset CreatedAt, int CreatedPlaylistsCount,
-    int PendingTracksCount, int ApprovedTracksCount);
+public record UserProfileResponse(
+    string Username,
+    string Email,
+    DateTimeOffset CreatedAt,
+    long CreatedPlaylistsCount,
+    long PendingTracksCount,
+    long ApprovedTracksCount);
